+++ test/acceptance/tests/hcp/terraform/hcp-install/variables.tf
@@ -48,5 +48,5 @@ variable "tags" {
 variable "consul_ecs_image" {
   description = "Consul ECS image to use."
   type        = string
-  default     = "docker.mirror.hashicorp.services/hashicorpdev/consul-ecs:latest"
+  default     = "docker.mirror.hashicorp.services/hashicorpdev/consul-ecs:9afa0c9"
 }